refactor(importer): gather the typed fact readers into facts.rs
Pure move, zero behavior change. Every helper whose job is reading a
fact from rustc through typed APIs now lives in one module,
translator/facts.rs, with the oracle contract in its doc:

  rustc (typed APIs) ──► facts.rs ──► the rest of the importer
  a fact is exact, or it is a hard error; never a guess, never a
  Debug string. New rustc questions go here, nowhere else.

Gathered: KnownDefs and its per-run cell, is_cuda_device_adt,
is_cuda_device_fn, is_std_precondition_check, self_ty_is_shared_array,
eval_usize_const (shared extraction of the SharedArray and
DynamicSharedArray const-generic readers, error strings unchanged),
extract_enum_variant. Translation machinery that merely consumes the
facts stays where it was.

Validated: full workspace suites, 218/218 examples compile and run on
sm_120a, gemm_views bench at parity, fmt, clippy.
