+++ crates/mir-importer/src/translator/terminator/mod.rs
@@ -60,6 +60,9 @@ pub mod intrinsics;
 
 use super::types;
 use crate::error::{TranslationErr, TranslationResult};
+use crate::translator::facts::{
+    eval_usize_const, is_cuda_device_fn, is_std_precondition_check, self_ty_is_shared_array,
+};
 use crate::translator::location::span_to_location;
 use crate::translator::rvalue;
 use crate::translator::values::{ValueMap, maybe_ptr_coerce};
@@ -882,21 +885,6 @@ fn is_drop_in_place_callee(fn_def: &rustc_public::ty::FnDef) -> bool {
     method == "drop_in_place" && parent == "ptr"
 }
 
-/// True when `fn_def` is a `precondition_check` shim generated by libcore's
-/// `assert_unsafe_precondition!` macro: a nested `const fn` literally named
-/// `precondition_check` inside the guarded function (e.g.
-/// `core::num::<impl usize>::unchecked_sub::precondition_check`). The macro
-/// expands in `core` and (for `Vec` internals) `alloc` — the same crates
-/// whose shim bodies the collector skips. Crate anchor + exact tail-segment
-/// match, so a user fn such as `my_precondition_check` stays a normal call.
-fn is_std_precondition_check(fn_def: &rustc_public::ty::FnDef) -> bool {
-    if !matches!(fn_def.krate().name.as_str(), "core" | "alloc") {
-        return false;
-    }
-    let name = fn_def.name();
-    name.as_str().rsplit("::").next() == Some("precondition_check")
-}
-
 /// Emit a branch to `target` as the only effect of a call we are eliding:
 /// the callee does nothing observable and has no device definition (a UB
 /// precondition check, or provably no-op drop glue). `emit_goto` needs a
@@ -992,9 +980,7 @@ fn translate_call(
     // Is the trait-method Self type SharedArray? Shared with
     // `values::classify_call` so intrinsic dispatch and destination-slot
     // classification can't drift.
-    let on_shared_array = substs_args
-        .as_ref()
-        .is_some_and(crate::translator::values::self_ty_is_shared_array);
+    let on_shared_array = substs_args.as_ref().is_some_and(self_ty_is_shared_array);
 
     // Skip precondition_check calls - these are UB check assertions that are
     // dead code because we return false for RuntimeChecks(UbChecks).
@@ -1156,14 +1142,7 @@ fn translate_call(
                 )
             );
         };
-        let alignment: u64 = c.eval_target_usize().map_err(|e| {
-            input_error!(
-                loc.clone(),
-                TranslationErr::unsupported(format!(
-                    "DynamicSharedArray ALIGN const generic did not evaluate to a target usize: {e:?}"
-                ))
-            )
-        })?;
+        let alignment: u64 = eval_usize_const(c, "DynamicSharedArray ALIGN", Some(&loc))?;
 
         if name.contains("::get") {
             // Both get() and get_raw() use the same handler with offset 0
@@ -2228,23 +2207,6 @@ fn is_cuda_device_const_marker(func: &mir::Operand, expected_name: &str) -> bool
     }
     let definition_name = definition.name();
     definition_name == expected_name || definition_name.ends_with(&format!("::{expected_name}"))
-}
-
-/// `true` if `func` is an fn item defined in the `cuda_device` crate.
-///
-/// Anchors name-substring dispatch gates (e.g. `DynamicSharedArray::get`)
-/// so a user fn merely spelled like a cuda-device one can't hijack an
-/// intrinsic lowering; it falls through to ordinary call handling instead.
-fn is_cuda_device_fn(func: &mir::Operand) -> bool {
-    use rustc_public::ty::{RigidTy, TyKind};
-
-    let mir::Operand::Constant(constant) = func else {
-        return false;
-    };
-    let TyKind::RigidTy(RigidTy::FnDef(definition, _)) = constant.const_.ty().kind() else {
-        return false;
-    };
-    definition.krate().name.as_str() == "cuda_device"
 }
 
 /// Extracts function metadata from a MIR function operand.
